fix: make one file to test cli in general
